Added LookStringLeft and LookStringRight. No tests yet. Needed for comment /* My comment */ Multi character symbols.

+++ sn/sni_thread.h
@@ -28,7 +28,7 @@ namespace SNI
 	{
 	public:
 		static SNI_Thread *GetThread();
-		static SNI_Thread *GetThreadByNumber(long p_ThreadNum);
+		static SNI_Thread *GetThreadByNumber(size_t p_ThreadNum);
 		static size_t GetNumThreads();
 		static SNI_Thread *BuildThread();
 		static SNI_Manager *TopManager();
